changed data type from linevalue purchasePirce to match Product.Price type, added test to check in penny format

+++ server/db/models/lineItem.js
@@ -3,7 +3,7 @@ const db = require('../db');
 
 const LineItem = db.define('lineItem', {
   purchasePrice: {
-    type: DECIMAL(12, 2),
+    type: INTEGER,
     allowNull: false
   },
   purchaseNum: {
